requirements.txt
pyaudio
Pillow
opencv-python
SpeechRecognition
matplotlib
BingImageCreator
undetected_chromedriver
google-generativeai
selenium
requests
pytube
pygame
keyboard
sounddevice
numpy
playaudio
playsound
webdriver_manager
g4f
pyperclip
pywhatkit
pyautogui
pyttsx3
html2text
gradio
simpleaudio
googlesearch-python
groq
psutil
edge-tts
pandas
webscout
appdirs
mtranslate
cohere
easyocr
transformers
accelerate
difflib